some cleanup of unnecessary things
